refactor: rename errors to error, add example into workspace
Rename to `error` to follow standard library conventions, and add example into workspace to share compilation cache in root `/target` folder.

Machine generated `Cargo.lock` file is merged into project root.
