+++ src/transport.rs
@@ -25,14 +25,14 @@ impl<I2C> SyncTransport for I2C
 where
     I2C: embedded_hal::i2c::I2c,
 {
-    type Error = crate::errors::TCA9534Error<I2C::Error>;
+    type Error = crate::error::TCA9534Error<I2C::Error>;
 
     fn write(&mut self, addr: u8, bytes: &[u8]) -> Result<(), Self::Error> {
-        I2C::write(self,addr, bytes).map_err(crate::errors::TCA9534Error::I2c)
+        I2C::write(self,addr, bytes).map_err(crate::error::TCA9534Error::I2c)
     }
 
     fn read(&mut self, addr: u8, bytes: &mut [u8]) -> Result<(), Self::Error> {
-        I2C::read(self,addr, bytes).map_err(crate::errors::TCA9534Error::I2c)
+        I2C::read(self,addr, bytes).map_err(crate::error::TCA9534Error::I2c)
     }
 
     fn write_read(
@@ -41,7 +41,7 @@ where
         wr_bytes: &[u8],
         rd_bytes: &mut [u8],
     ) -> Result<(), Self::Error> {
-        I2C::write_read(self,addr, wr_bytes, rd_bytes).map_err(crate::errors::TCA9534Error::I2c)
+        I2C::write_read(self,addr, wr_bytes, rd_bytes).map_err(crate::error::TCA9534Error::I2c)
     }
 }
 
@@ -72,14 +72,14 @@ impl<I2C> AsyncTransport for I2C
 where
     I2C: embedded_hal_async::i2c::I2c,
 {
-    type Error = crate::errors::TCA9534Error<I2C::Error>;
+    type Error = crate::error::TCA9534Error<I2C::Error>;
 
     async fn write(&mut self, addr: u8, bytes: &[u8]) -> Result<(), Self::Error> {
-        I2C::write(self,addr, bytes).await.map_err(crate::errors::TCA9534Error::I2c)
+        I2C::write(self,addr, bytes).await.map_err(crate::error::TCA9534Error::I2c)
     }
 
     async fn read(&mut self, addr: u8, bytes: &mut [u8]) -> Result<(), Self::Error> {
-        I2C::read(self,addr, bytes).await.map_err(crate::errors::TCA9534Error::I2c)
+        I2C::read(self,addr, bytes).await.map_err(crate::error::TCA9534Error::I2c)
     }
 
     async fn write_read(
@@ -88,7 +88,7 @@ where
         wr_bytes: &[u8],
         rd_bytes: &mut [u8],
     ) -> Result<(), Self::Error> {
-        I2C::write_read(self,addr, wr_bytes, rd_bytes).await.map_err(crate::errors::TCA9534Error::I2c)
+        I2C::write_read(self,addr, wr_bytes, rd_bytes).await.map_err(crate::error::TCA9534Error::I2c)
     }
 }
